requirements.txt
aiofiles
aiohttp
bs4
covid
cowpy
dnspython
emoji
gitpython
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googletrans==3.1.0a0
hachoir
heroku3
html-telegraph-poster
instaloader
motor
natsort
oauth2client
Pillow==8.1.2
psutil
pybase64
pyrogram==1.1.12
pySmartDL
python-dotenv
pytz
rarfile
removebg
requests
search-engine-parser
selenium==3.141.0
setuptools>=40.3.0
spamwatch
speedtest-cli
stagger
telegraph
tgcrypto
urbandict==0.5
wget
wikipedia
youtube_dl>=2021.1.8
